requirements.txt
Werkzeug==1.0.1
Flask_CKEditor==0.4.6
Flask_Script==2.0.6
Flask==1.1.2
cassandra_driver==3.25.0

cassandra-driver~=3.25.0
requests~=2.25.1
